COM-1225 rename common.js
