+++ src/components/form/Button.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { View, TouchableOpacity, Text, StyleSheet } from 'react-native';
-import commonStyle from '../../styles/common';
+import variables from '../../styles/variables';
 
 const NiButton = ({ style, caption, onPress }) => {
   return (
@@ -15,7 +15,7 @@ const NiButton = ({ style, caption, onPress }) => {
 const styles = StyleSheet.create({
   container: {},
   button: {
-    backgroundColor: commonStyle.PRIMARY_COLOR,
+    backgroundColor: variables.PRIMARY_COLOR,
     borderRadius: 2,
     display: 'flex',
     flexDirection: 'row',
